[6.17] fixed bug where ingredients of recipe were multiplied incorrectly

+++ backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/endpoint/ShoppingListEndpoint.java
@@ -154,7 +154,7 @@ public class ShoppingListEndpoint {
     @PutMapping
     @Operation(summary = "Plan a recipe: adds missing ingredients to shoppingList", security = @SecurityRequirement(name = "apiKey"))
     public List<ItemStorageDto> planRecipe(Authentication authentication,
-                                           @RequestParam(name = "recipeId") Long recipeId, @RequestParam(name = "people") Long people) {
+                                           @RequestParam(name = "recipeId") Long recipeId, @RequestParam(name = "people") Integer people) {
         LOGGER.info("Endpoint: POST /api/v1/shoppinglist/recipeId={}, people={},userName={}", recipeId, people, authentication.getName());
         try {
             return itemStorageMapper.itemsStorageToItemsStorageDto(
@@ -173,7 +173,7 @@ public class ShoppingListEndpoint {
     @PutMapping(value = "/putAllIngredientsOfRecipe")
     @Operation(summary = "Adds all ingredients of a recipe to shoppingList", security = @SecurityRequirement(name = "apiKey"))
     public List<ItemStorageDto> putRecipeOnShoppingList(Authentication authentication,
-                                                        @RequestParam(name = "recipeId") Long recipeId, @RequestParam(name = "people") Long people) {
+                                                        @RequestParam(name = "recipeId") Long recipeId, @RequestParam(name = "people") Integer people) {
         LOGGER.info("Endpoint: POST /api/v1/shoppinglist/putAllIngredientsOfRecipe/recipeId={},people={},userName={}", recipeId, people, authentication.getName());
         try {
             return itemStorageMapper.itemsStorageToItemsStorageDto(
